refactor: move steam parser to utils
- Updated import paths in `steam-guide-card.tsx` to utilize the new utility, improving code organization and maintainability.
