+++ src/components/steam-guide/steam-guide-card.tsx
@@ -1,11 +1,11 @@
 import Image from "next/image";
+import { parseDescription } from "@/utils/parse-steam-guide";
 import { ArrowRightIcon, BookHeartIcon, EyeIcon, StarIcon } from "lucide-react";
 import millify from "millify";
 
 import type { SteamGuideResponse } from "@/types/steam";
 import { getApiUrl } from "@/lib/api";
 import { steamConfig } from "@/lib/constants";
-import { parseDescription } from "@/lib/parseSteamGuide";
 import { Button } from "@/components/ui/button";
 import {
   Card,
